fix(mcp/fixtures): split fixture tools to stay under 200-line lint limit
Extract write tools (create, update, delete, connect, disconnect) to
inventory-fixtures-write.ts (134 lines); read tools (list, get,
listForItem) remain in inventory-fixtures.ts (78 lines).
fixtureTools barrel unchanged — still exports all 8 tools.
